Add look text for charges Fix charges being decreased on non protection damage
