+++ src/Game/NeoServer.Game.Common/Contracts/Items/Types/IProtection.cs
@@ -4,6 +4,6 @@ namespace NeoServer.Game.Common.Contracts.Items.Types
 {
     public interface IProtection
     {
-        void Protect(ref CombatDamage damage);
+        bool Protect(ref CombatDamage damage);
     }
 }
